+++ source/Messaging.Infrastructure/Configuration/CompositionRoot.cs
@@ -61,7 +61,7 @@ using Messaging.Infrastructure.MasterData.MarketEvaluationPoints;
 using Messaging.Infrastructure.OutgoingMessages;
 using Messaging.Infrastructure.OutgoingMessages.AccountingPointCharacteristics;
 using Messaging.Infrastructure.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
-using Messaging.Infrastructure.OutgoingMessages.ConfirmChangeAccountingPointCharacteristics;
+using Messaging.Infrastructure.OutgoingMessages.ConfirmRequestChangeAccountingPointCharacteristics;
 using Messaging.Infrastructure.OutgoingMessages.ConfirmRequestChangeOfSupplier;
 using Messaging.Infrastructure.OutgoingMessages.GenericNotification;
 using Messaging.Infrastructure.OutgoingMessages.RejectRequestChangeOfSupplier;
@@ -284,7 +284,7 @@ namespace Messaging.Infrastructure.Configuration
             _services.AddScoped<IDocumentWriter, RejectRequestChangeOfSupplierXmlDocumentWriter>();
             _services.AddScoped<IDocumentWriter, GenericNotificationDocumentWriter>();
             _services.AddScoped<IDocumentWriter, AccountingPointCharacteristicsDocumentWriter>();
-            _services.AddScoped<IDocumentWriter, ConfirmChangeAccountingPointCharacteristicsDocumentWriter>();
+            _services.AddScoped<IDocumentWriter, ConfirmRequestChangeAccountingPointCharacteristicsDocumentWriter>();
             _services.AddScoped<IDocumentWriter, CharacteristicsOfACustomerAtAnApDocumentWriter>();
             _services.AddScoped<IDocumentWriter, RejectRequestChangeOfSupplierJsonDocumentWriter>();
             _services.AddScoped<IValidationErrorTranslator, ValidationErrorTranslator>();
